Put files into different folders by functionality
This commit is a preparation for supporting SWR.

* Move common modules which are used by React-Query and SWR into `/common `
* Move the rest modules into `react-query`
